Add close button to mission success screen
On the "mission success" screen, the only available choice is worded the same way as it was in the original game, "You have completed the mission." That this button actually leads to the Author's Note for Schrödinger's Door is not at all clear.

In Tau Station, you would normally go do something else after finishing a mission. Since Schrödinger's Door is standalone, users may not be looking for a way to get rid of the mission success screen other than closing the browser tab.

Perhaps adding an X-style close button similar to what Tau Station itself used might help some users.